# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Cohort — AI agent fleet governance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /discovery/runs:
    post:
      operationId: startDiscoveryRun
      tags: [agents]
      summary: Start a mass-discovery run
      description: >-
        Opens a discovery run that groups the agents found from a source
        (connector platform, GitHub org, etc.) and stages them as editable
        drafts before any admission. The rich rule-based / AI enrichment is
        performed by separate workers; this endpoint persists the run and the
        bare staged drafts.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartDiscoveryRunInput"
      responses:
        "201":
          description: Discovery run started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoveryRun"
  /discovery/runs/{runId}:
    get:
      operationId: getDiscoveryRun
      tags: [agents]
      summary: Get a discovery run with progress counts
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Discovery run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoveryRun"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /discovery/drafts:
    get:
      operationId: listAgentDrafts
      tags: [agents]
      summary: List staged agent drafts with filters
      parameters:
        - name: runId
          in: query
          required: false
          schema:
            type: string
        - name: reviewStatus
          in: query
          required: false
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: enrichmentStatus
          in: query
          required: false
          schema:
            type: string
            enum: [pending, rules, enriching, enriched, failed]
        - name: platform
          in: query
          required: false
          schema:
            type: string
        - name: minConfidence
          in: query
          required: false
          schema:
            type: number
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Staged drafts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentDraftRecord"
  /discovery/drafts/bulk-approve:
    post:
      operationId: bulkApproveAgentDrafts
      tags: [agents]
      summary: Approve multiple drafts (promote to fleet)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkDraftIdsInput"
      responses:
        "200":
          description: Bulk approval result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkReviewResult"
  /discovery/drafts/bulk-reject:
    post:
      operationId: bulkRejectAgentDrafts
      tags: [agents]
      summary: Reject multiple drafts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkDraftIdsInput"
      responses:
        "200":
          description: Bulk rejection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkReviewResult"
  /discovery/drafts/{draftId}:
    patch:
      operationId: updateAgentDraft
      tags: [agents]
      summary: Edit a staged draft before review
      parameters:
        - name: draftId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentDraftUpdate"
      responses:
        "200":
          description: Updated draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDraftRecord"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /discovery/drafts/{draftId}/approve:
    post:
      operationId: approveAgentDraft
      tags: [agents]
      summary: Approve a draft and promote it to the fleet
      description: >-
        Promotes the draft into the real fleet reusing the standard agent
        admission path (Carteira de Trabalho + seeded initial evaluation), then
        marks the draft as approved and links it to the created agent.
      parameters:
        - name: draftId
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Agent admitted from draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Draft already reviewed or agent already admitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /discovery/drafts/{draftId}/reject:
    post:
      operationId: rejectAgentDraft
      tags: [agents]
      summary: Reject a draft without touching the fleet
      parameters:
        - name: draftId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectDraftInput"
      responses:
        "200":
          description: Rejected draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDraftRecord"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Agent:
      type: object
      required:
        - id
        - name
        - slug
        - role
        - platform
        - version
        - status
        - bio
        - currentVerdict
        - verdictConfidence
        - severity
        - healthScore
        - admittedAt
        - lastEvaluatedAt
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        role:
          type: string
        platform:
          type: string
        version:
          type: string
        status:
          type: string
          enum: [observation, active, flagged, retiring, retired]
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: string
        tagline:
          type: string
        monthlyVolume:
          type: integer
        headlineKpis:
          type: array
          items:
            $ref: "#/components/schemas/KpiMetric"
        targetMetrics:
          type: array
          description: >-
            Flattened metrics from the agent's latest evaluation that carry a
            comparable goal. Used by the fleet view to roll up how many metrics
            are currently off-target.
          items:
            $ref: "#/components/schemas/KpiMetric"
        currentVerdict:
          type: string
          enum: [promote, mentor, retire, observation]
        verdictConfidence:
          type: number
        severity:
          type: string
          enum: [critical, high, medium, stable]
        healthScore:
          type: number
        activeAlerts:
          type: integer
        monthlyValue:
          type: number
        monthlyCost:
          type: number
        admittedAt:
          type: string
        lastEvaluatedAt:
          type: string
    AgentDetail:
      type: object
      required:
        - agent
        - identity
        - owners
        - latestEvaluation
      properties:
        agent:
          $ref: "#/components/schemas/Agent"
        identity:
          $ref: "#/components/schemas/AgentIdentity"
        owners:
          $ref: "#/components/schemas/AgentOwners"
        latestEvaluation:
          $ref: "#/components/schemas/Evaluation"
        currentVerdict:
          $ref: "#/components/schemas/Verdict"
    AgentIdentity:
      type: object
      required:
        - bio
        - shouldDo
        - shouldNotDo
        - autonomyLevel
        - limits
        - businessCase
        - version
      properties:
        bio:
          type: string
        shouldDo:
          type: array
          items:
            type: string
        shouldNotDo:
          type: array
          items:
            type: string
        autonomyLevel:
          type: string
          enum: [autonomous, escalates, restricted]
        autonomyNotes:
          type: string
        limits:
          type: array
          items:
            type: string
        businessCase:
          $ref: "#/components/schemas/BusinessCase"
        version:
          type: integer
    BusinessCase:
      type: object
      required:
        - baseline
        - targetPayback
        - actualPayback
        - description
      properties:
        baseline:
          type: string
        targetPayback:
          type: string
        actualPayback:
          type: string
        description:
          type: string
    AgentOwners:
      type: object
      required:
        - businessOwner
        - technicalOwner
        - governanceSponsor
      properties:
        businessOwner:
          type: string
        technicalOwner:
          type: string
        governanceSponsor:
          type: string
    KpiMetric:
      type: object
      required:
        - label
        - value
        - unit
        - trend
      properties:
        label:
          type: string
        value:
          type: number
        unit:
          type: string
        trend:
          type: number
        direction:
          type: string
          enum: [up, down, flat]
        target:
          type: string
        rationale:
          type: string
    KpiLayer:
      type: object
      required:
        - key
        - label
        - score
        - severity
        - metrics
      properties:
        key:
          type: string
          enum: [efficacy, efficiency, adoption, governance, value]
        label:
          type: string
        score:
          type: number
        severity:
          type: string
          enum: [critical, high, medium, stable]
        metrics:
          type: array
          items:
            $ref: "#/components/schemas/KpiMetric"
...
    Evaluation:
      type: object
      required:
        - id
        - agentId
        - evaluatedAt
        - window
        - layers
        - verdict
        - verdictConfidence
        - rationale
      properties:
        id:
          type: string
        agentId:
          type: string
        evaluatedAt:
          type: string
        window:
          type: string
        layers:
          type: array
          items:
            $ref: "#/components/schemas/KpiLayer"
        verdict:
          type: string
          enum: [promote, mentor, retire, observation]
        verdictConfidence:
          type: number
        rationale:
          type: string
...
    NextAction:
      type: object
      required:
        - action
        - owner
        - due
      properties:
        action:
          type: string
        owner:
          type: string
        due:
          type: string
    Verdict:
      type: object
      required:
        - id
        - agentId
        - verdict
        - confidence
        - executionWindow
        - suggestedSponsor
        - nextActions
        - rationale
        - decision
        - createdAt
      properties:
        id:
          type: string
        agentId:
          type: string
        verdict:
          type: string
          enum: [promote, mentor, retire, observation]
        confidence:
          type: number
        executionWindow:
          type: string
        suggestedSponsor:
          type: string
        nextActions:
          type: array
          items:
            $ref: "#/components/schemas/NextAction"
        rationale:
          type: string
        decision:
          type: string
          enum: [pending, approved, disagreed, exported]
        decidedBy:
          type: ["string", "null"]
        decidedAt:
          type: ["string", "null"]
        createdAt:
          type: string
...
    DraftMetric:
      type: object
      required:
        - layer
        - label
        - unit
        - target
      properties:
        layer:
          type: string
          enum: [efficacy, efficiency, adoption, governance, value]
        label:
          type: string
        unit:
          type: string
        target:
          type: string
        value:
          type: number
          description: >-
            Optional reviewer-set starting/current value for the metric. When
            provided during admission it overrides the deterministically seeded
            value so goal-vs-actual reflects reality.
        rationale:
          type: string
    DraftBusinessCase:
      type: object
      required:
        - baseline
        - targetPayback
        - description
      properties:
        baseline:
          type: string
        targetPayback:
          type: string
        description:
          type: string
...
    StartDiscoveryRunInput:
      type: object
      required:
        - source
      properties:
        source:
          type: string
          description: Origin kind, e.g. "connector", "github", "manual".
        sourceRef:
          type: string
          description: >-
            Source reference — e.g. a platform key for a connector, or a GitHub
            org/repo. For connector sources matching the platform catalog, the
            run stages a bare draft per discovered agent.
        note:
          type: string
    DiscoveryRunCounts:
      type: object
      required:
        - total
        - pendingReview
        - approved
        - rejected
        - enrichmentPending
        - enrichmentRules
        - enriching
        - enriched
        - enrichmentFailed
      properties:
        total:
          type: integer
        pendingReview:
          type: integer
        approved:
          type: integer
        rejected:
          type: integer
        enrichmentPending:
          type: integer
        enrichmentRules:
          type: integer
        enriching:
          type: integer
        enriched:
          type: integer
        enrichmentFailed:
          type: integer
    DiscoveryRun:
      type: object
      required:
        - id
        - source
        - status
        - totalDiscovered
        - draftsCreated
        - note
        - startedAt
        - createdAt
        - counts
      properties:
        id:
          type: string
        source:
          type: string
        sourceRef:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, running, completed, failed]
        totalDiscovered:
          type: integer
        draftsCreated:
          type: integer
        note:
          type: string
        startedAt:
          type: string
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        counts:
          $ref: "#/components/schemas/DiscoveryRunCounts"
    AgentDraftRecord:
      type: object
      required:
        - id
        - runId
        - source
        - name
        - role
        - platform
        - tagline
        - bio
        - shouldDo
        - shouldNotDo
        - autonomyLevel
        - limits
        - businessCase
        - proposedMetrics
        - summary
        - confidence
        - enrichmentStatus
        - reviewStatus
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        runId:
          type: string
        source:
          type: string
        externalId:
          type: ["string", "null"]
        name:
          type: string
        role:
          type: string
        platform:
          type: string
        tagline:
          type: string
        bio:
          type: string
        shouldDo:
          type: array
          items:
            type: string
        shouldNotDo:
          type: array
          items:
            type: string
        autonomyLevel:
          type: string
          enum: [autonomous, escalates, restricted]
        autonomyNotes:
          type: ["string", "null"]
        limits:
          type: array
          items:
            type: string
        businessCase:
          $ref: "#/components/schemas/DraftBusinessCase"
        proposedMetrics:
          type: array
          items:
            $ref: "#/components/schemas/DraftMetric"
        summary:
          type: string
        confidence:
          type: number
        enrichmentStatus:
          type: string
          enum: [pending, rules, enriching, enriched, failed]
        reviewStatus:
          type: string
          enum: [pending, approved, rejected]
        promotedAgentId:
          type: ["string", "null"]
        reviewNote:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    AgentDraftUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        platform:
          type: string
        tagline:
          type: string
        bio:
          type: string
        shouldDo:
          type: array
          items:
            type: string
        shouldNotDo:
          type: array
          items:
            type: string
        autonomyLevel:
          type: string
          enum: [autonomous, escalates, restricted]
        autonomyNotes:
          type: string
        limits:
          type: array
          items:
            type: string
        businessCase:
          $ref: "#/components/schemas/DraftBusinessCase"
        proposedMetrics:
          type: array
          items:
            $ref: "#/components/schemas/DraftMetric"
        summary:
          type: string
        confidence:
          type: number
    RejectDraftInput:
      type: object
      properties:
        reason:
          type: string
    BulkDraftIdsInput:
      type: object
      required:
        - draftIds
      properties:
        draftIds:
          type: array
          items:
            type: string
        reason:
          type: string
    BulkReviewResult:
      type: object
      required:
        - requested
        - succeeded
        - failed
        - agentIds
        - draftIds
      properties:
        requested:
          type: integer
        succeeded:
          type: integer
        failed:
          type: integer
        agentIds:
          type: array
          items:
            type: string
        draftIds:
          type: array
          items:
            type: string
...